Ported in changes from COTCAC by Jele155,  released on 8/10/21 and last updated on 10/17/21
This is why I forked from TehCJay-master instead of from one of the more recent GG forks, because this fork will use COTCAC as a starting point, and as far as I can tell COTAC is based on master but one commit behind so that it is near to it's correct position in the git history. I chose to start from the COTAC version because it is the least buggy version by currently even though it is missing many of our newer features and reworks. I intend to import additional mods that do not have many conflicting files to resolve societies such as Orders of Chivalry for warrior societies, EMF for merchant society, perhaps Fantasy Invasion for their secret societies and invasions, and Ancient Religions Reborn and expanded to hopefully enable pagan reformation, the purpose of this fork is primarily so I can keep track of what these import attempts change so I can resolve any file conflicts to make them compatible with this new COTCAC+ version.