WIP files for Terran Universalism ideology and society adapted from the Communism mod by Hecthec
Made separate branch because there are noculture characters spawning from this in large numbers, and I wasn't able to locate why in the events. Localization was updated for the religion/ideology itself but not yet for the society. There are likely other unidentified issues, hoping to return to implementing this into the master branch in a few weeks. There seemed to be a new source of lag so trying to ensure it was caused by this and not other recent additions.